producer/requirements.txt
Faker==10.0.0
SQLAlchemy==1.4.28
psycopg2-binary==2.9.2
google-cloud-pubsub==2.9.0